[FEATURE] Symfony2 Bundle
Phase 1 complete: basic template support.

Enables the use of Fluid as a Symfony2 package which provides an additional rendering engine named "fluid". Install the Bundle, add the "fluid" engine to the list of rendering engines and use the Symfony2 naming convention for your template files, i.e.:

@MyPackage/Resources/views/Controller/Action.format.engine
@MyPackage/Resources/views/Controller/Index.html.fluid

Phase 1 does not support the use of template paths and thus not the use of Layouts and Partials. This will be added in Phase 2.